requirements-dev.txt
# Dev/test dependencies
pytest>=8.0
pyyaml>=6.0
